+++ src/components/dialog.js
@@ -1,19 +1,19 @@
 import React, { useState, useEffect } from "react";
 import {
-    View,
-    Text,
-    Button,
-    ImageBackground,
-    Image,
-    TouchableOpacity,
-    ScrollView,
-  } from "react-native";
+  View,
+  Text,
+  Button,
+  ImageBackground,
+  Image,
+  TouchableOpacity,
+  ScrollView,
+} from "react-native";
 
-  export default DialogBox = ({ message }) => {
-    return (
-      <View style={message.style}>
-        {/* change text color */}
-        <Text>{message.text}</Text>
-      </View>
-    );
-  };
+export default DialogBox = ({ message }) => {
+  return (
+    <View style={message.style}>
+      {/* change text color */}
+      <Text style={{ color: "#FFFFFF" }}>{message.text}</Text>
+    </View>
+  );
+};
